Unstable update
This is untested!
* worked on new implementation of the applications
* cancel should be working now
* config can now create a config file (will get a text UI in the future)
* process can now execute external programs
* improved slurm interface
* some bug fixes

new test scheme:
test.test.TESTMODE will define the behavior of many functions.
If true file IO and external calls will be omitted,
instead debug results will be provided

+++ sutils/parameters/numeric.py
@@ -18,4 +18,4 @@ def step_range(lo, hi, step):
     if is_int(lo) and is_int(hi) and is_int(step):
         return np.arange(lo, hi+step, step, dtype=int)
     else:
-        return np.arange(lo, hi+step, step, dtype=int)
+        return np.arange(lo, hi+step, step)
